Modify toggleScaleDownModal function
