+++ shell/models/cluster.x-k8s.io.machinedeployment.js
@@ -6,8 +6,8 @@ import { exceptionToErrorsArray } from '@shell/utils/error';
 import { handleConflict } from '@shell/plugins/dashboard-store/normalize';
 import { MACHINE_ROLES } from '@shell/config/labels-annotations';
 import { notOnlyOfRole } from '@shell/models/cluster.x-k8s.io.machine';
-import { isAlternate } from '@shell/utils/platform';
-import { SCALE_POOL_PROMPT } from '@shell/store/prefs';
+// import { isAlternate } from '@shell/utils/platform';
+// import { SCALE_POOL_PROMPT } from '@shell/store/prefs';
 
 export default class CapiMachineDeployment extends SteveModel {
   get cluster() {
@@ -102,27 +102,32 @@ export default class CapiMachineDeployment extends SteveModel {
     return machinePools.find(pool => pool.machineConfigRef.name === machineConfigName);
   }
 
-  toggleScaleDownModal( event, resources = this ) {
-    // Check if the user held alt key when an action is clicked.
-    const alt = isAlternate(event);
-    const showScalePoolPrompt = this.$rootGetters['prefs/get'](SCALE_POOL_PROMPT);
+  // toggleScaleDownModal( event, resources = this ) {
+  //   // Check if the user held alt key when an action is clicked.
+  //   const alt = isAlternate(event);
+  //   const showScalePoolPrompt = this.$rootGetters['prefs/get'](SCALE_POOL_PROMPT);
 
-    // Prompt if showScalePoolPrompt pref not store and user did not held alt key
-    if (!alt && !showScalePoolPrompt) {
-      this.$dispatch('promptModal', {
-        component:  'ScalePoolDownDialog',
-        resources,
-        modalWidth: '450px'
-      });
-    } else {
-      // User held alt key, so don't prompt
-      this.scalePool(-1);
-    }
-  }
+  //   // Prompt if showScalePoolPrompt pref not store and user did not held alt key
+  //   if (!alt && !showScalePoolPrompt) {
+  //     this.$dispatch('promptModal', {
+  //       component:  'ScalePoolDownDialog',
+  //       resources,
+  //       modalWidth: '450px'
+  //     });
+  //   } else {
+  //     // User held alt key, so don't prompt
+  //     this.scalePool(-1);
+  //   }
+  // }
 
   scalePool(delta, save = true, depth = 0) {
     // This is used in different places with different scaling rules, so don't check if we can/cannot scale
     if (!this.inClusterSpec) {
+      return;
+    }
+    if ( delta === -1) {
+      console.log('-1');
+
       return;
     }
